% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\docType{package}
\name{actyR-package}
\alias{actyR-package}
\title{R / C++ programs for Abbring, Campbell, Tilly Yang (2017) Very Simple Markov
Perfect Industry Dynamics: Empirics}
\description{
R / C++ programs for Abbring, Campbell, Tilly Yang (2017) Very Simple Markov
Perfect Industry Dynamics: Empirics
}
\seealso{
A Matlab sandbox is
  available at \url{https://github.com/jtilly/acty}. An extensive
  documentation of the Matlab sandbox is available at
  \url{https://jtilly.github.io/acty}.
}
\author{
Jan Tilly
}
