% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeMixingDensity}
\alias{computeMixingDensity}
\title{Compute the mixing density}
\usage{
computeMixingDensity(Settings, Param, vS)
}
\arguments{
\item{Settings}{is an R list of settings}

\item{Param}{is an R list of of parameters}

\item{vS}{is a matrix with the value functions}
}
\value{
A cube with the mixing density.
}
\description{
This function computes the mixing density and returns it as a cube.
}
