% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\docType{data}
\name{cov}
\alias{cov}
\title{Covariates for muSAs in the year 2000}
\format{A data frame with 573 rows and 12 variables
\describe{
  \item{id}{CBSA id}
  \item{name}{CBSA name}
  \item{inc}{median income in 2000}
  \item{dummy1}{Middle Atlantic Census Region Dummy}
  \item{dummy2}{East North Central Census Region Dummy}
  \item{dummy3}{West North Central Census Region Dummy}
  \item{dummy4}{South Atlantic Census Region Dummy}
  \item{dummy5}{East South Central Census Region Dummy}
  \item{dummy6}{West South Central Census Region Dummy}
  \item{dummy7}{Mountain Census Region Dummy}
  \item{dummy8}{Pacific Census Region Dummy}
  \item{dummy9}{Low Diversity Dummy}
}}
\source{
\url{http://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml}
}
\description{
This data set contains covariates for the muSAs in the year 2000
}
