% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dPidMu}
\alias{dPidMu}
\title{Compute the gradient of the transition probability matrix with respect to mu}
\usage{
dPidMu(Pi, logGrid, mu, sigma)
}
\arguments{
\item{Pi}{is the transition probability matrix}

\item{logGrid}{is the demand grid}

\item{mu}{is the mean of the innovations times 100}

\item{sigma}{is the standard deviation of the innovations times 100}
}
\value{
A matrix with the derivative of the transition probability matrix with respect to mu.
}
\description{
Compute the gradient of the transition probability matrix with respect to mu
}
