% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dPidSigma}
\alias{dPidSigma}
\title{Compute the gradient of the transition probability matrix with respect to sigma}
\usage{
dPidSigma(Pi, logGrid, mu, sigma)
}
\arguments{
\item{Pi}{is the transition probability matrix}

\item{logGrid}{is the demand grid}

\item{mu}{is the mean of the innovations}

\item{sigma}{is the standard deviation of the innovations times 100}
}
\value{
A matrix with the derivative of the transition probability matrix with respect to sigma.
}
\description{
Compute the gradient of the transition probability matrix with respect to sigma
}
