% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauchen.R
\name{ergodic}
\alias{ergodic}
\title{Compute the ergodic distribution of a Markov Chain}
\usage{
ergodic(transMat)
}
\arguments{
\item{transMat}{the transition probability matrix of the Markov chain}
}
\value{
a vector with the ergodic distribution
}
\description{
Computes the ergodic distribution of a discrete Markov chain
}
