% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\docType{data}
\name{firms}
\alias{firms}
\title{Firm count data for muSAs from 2000 to 2009}
\format{A data frame with 573 rows and 12 variables
\describe{
  \item{id}{CBSA id}
  \item{name}{CBSA name}
  \item{firms2000}{firms in 2000}
  \item{firms2001}{firms in 2001}
  \item{firms2002}{firms in 2002}
  \item{firms2003}{firms in 2003}
  \item{firms2004}{firms in 2004}
  \item{firms2005}{firms in 2005}
  \item{firms2006}{firms in 2006}
  \item{firms2007}{firms in 2007}
  \item{firms2008}{firms in 2008}
  \item{firms2009}{firms in 2009}
}}
\source{
\url{http://www.census.gov/econ/cbp/}
}
\description{
This data set contains firm counts for Movie theaters for muSAs from 2000 to
2009 in panel data format
}
