% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getModelTransitionMatrix}
\alias{getModelTransitionMatrix}
\title{Computes the model's transition matrix}
\usage{
getModelTransitionMatrix(Settings, Param)
}
\arguments{
\item{Settings}{is an R list with settings}

\item{Param}{is an R list with parameters}
}
\value{
square matrix with cCheck * (nCheck + 1) rows and columns
}
\description{
Computes the model's implied transition matrix
}
