% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood}
\alias{likelihood}
\title{Compute likelihood contributions}
\usage{
likelihood(Settings, Param, Data)
}
\arguments{
\item{Settings}{is an R list of settings}

\item{Param}{is an R list of parameters}

\item{Data}{is an R list of the data and includes the matrices \code{N} and \code{C}}
}
\value{
A column vector of likelihood contributions for each observation
}
\description{
This function computes the vector of likelihood contributions for the
data set in \code{Data}. The likelihood contributions are computed as follows:
\itemize{
\item{ Use the function \code{valuefunction} to compute the equilibrium value functions
and the entry and sure survival probabilities}
\item{Compute the mixing density}
\item{Assemble the likelihood contributions for each observations}
}
}
