% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihoodGradient}
\alias{likelihoodGradient}
\title{Compute the gradient of the likelihood function}
\usage{
likelihoodGradient(Settings, Param, Data)
}
\arguments{
\item{Settings}{is an R list of settings}

\item{Param}{is an R list of parameters}

\item{Data}{an R list of the data with matrices \code{C} and \code{N}}
}
\value{
An R list with a vector of likelihood contributions and a matrix of
gradient contributions.
}
\description{
This function computes the gradient of the likelihood function.
It also computes the equilibrium and the likelihood function and
then returns a list with the likelihood and gradient contributions.
}
