% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixingDensityGradient}
\alias{mixingDensityGradient}
\title{Compute the gradient of the mixing density for a variable of interest}
\usage{
mixingDensityGradient(Settings, Param, Eq, D_vS, OMEGA_FLAG = FALSE)
}
\arguments{
\item{Settings}{is an R list of settings}

\item{Param}{is an R list of parameters}

\item{Eq}{is an R list of equilibrium objects}

\item{D_vS}{is a matrix of the gradient of the value function with respect to the variable of interest}

\item{OMEGA_FLAG}{is a flag whether the variable in question is omega or not}
}
\value{
A cube of the mixing density \code{p^{ - 1}(a_S, c, n)}
}
\description{
This function computes the gradient of the mixing density. Note that all
parameters (except for omega) enter the mixing density only through the value functions. The
derivative of the value function is given as argument to the function.
}
