% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mixingProbabilities}
\alias{mixingProbabilities}
\title{Compute mixing probabilities}
\usage{
mixingProbabilities(N, C, W, vS)
}
\arguments{
\item{N}{is an integer with the number of firms post-entry}

\item{C}{is an integer with the index of the demand grid}

\item{W}{is the a scalar with the current realization of the cost shock}

\item{vS}{is the post-survival value function}
}
\value{
a scalar with the mixing probability
}
\description{
The function computes the mixing probability for a given state (N,C,W).
}
