% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\name{nfxpCovarianceStep3}
\alias{nfxpCovarianceStep3}
\title{Third Step Variance-Covariance Matrix}
\usage{
nfxpCovarianceStep3(x, Data, Settings, Param)
}
\arguments{
\item{x}{is a vector of parameters}

\item{Data}{is an R list that contains a matrix \code{C} with indices
of the demand state and a matrix \code{N} with number of active firms}

\item{Settings}{is an R list with settings}

\item{Param}{is an R list with parameters}
}
\value{
Returns a vector with likelihood contributions
}
\description{
This function computes the variance-covariance matrix that corresponds
to the third-step likelihood. The function uses the outer-product-of-the-gradient
estimator of the hessian.
}
