% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\name{nfxpGradientStep2}
\alias{nfxpGradientStep2}
\title{Second Step Gradient Function}
\usage{
nfxpGradientStep2(x, Data, Settings, Param)
}
\arguments{
\item{x}{is a vector of parameters}

\item{Data}{is an R list that contains a matrix \code{C} with indices
of the demand state and a matrix \code{N} with number of active firms}

\item{Settings}{is an R list with settings}

\item{Param}{is an R list with parameters}
}
\value{
Returns a vector of gradients
}
\description{
This function computes the gradient of the second step likelihood function
}
