% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\name{nfxpGradientStep3}
\alias{nfxpGradientStep3}
\title{Second Step Gradient Function}
\usage{
nfxpGradientStep3(x, Data, Settings, Param, Pi = NULL)
}
\arguments{
\item{x}{is a vector of parameters}

\item{Data}{is an R list that contains a matrix \code{C} with indices
of the demand state and a matrix \code{N} with number of active firms}

\item{Settings}{is an R list with settings}

\item{Param}{is an R list with parameters}

\item{Pi}{is the transition probabillity matrix if \code{Pi} is not
provided the transition probability matrix is computed based on \code{x}
times length(myGrid).}
}
\value{
Returns a vector of gradients
}
\description{
This function computes the gradient of the second step likelihood function
}
