% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\name{nfxpLikelihoodStep1}
\alias{nfxpLikelihoodStep1}
\title{First Step Likelihood Function}
\usage{
nfxpLikelihoodStep1(mu_sigma, Data, Settings, Pi = NULL)
}
\arguments{
\item{mu_sigma}{is a vector whose elements are \code{mu} and \code{sigma}}

\item{Data}{is an R list that contains a matrix \code{C} with indices
of the demand state}

\item{Settings}{is an R list with program Settings, most notably cCheck}

\item{Pi}{is the transition probabillity matrix if \code{Pi} is not
provided the transition probability matrix is computed based on \code{x}
times length(myGrid).}
}
\value{
Returns the scalar valued negative log-likelihood
}
\description{
This function computes the likelihood used in the first step of
the estimation procedure.
}
