% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfxp.R
\docType{data}
\name{population}
\alias{population}
\title{Population data for muSAs from 2000 to 2009}
\format{A data frame with 573 rows and 12 variables
\describe{
  \item{id}{CBSA id}
  \item{name}{CBSA name}
  \item{pop2000}{population in 2000}
  \item{pop2001}{population in 2001}
  \item{pop2002}{population in 2002}
  \item{pop2003}{population in 2003}
  \item{pop2004}{population in 2004}
  \item{pop2005}{population in 2005}
  \item{pop2006}{population in 2006}
  \item{pop2007}{population in 2007}
  \item{pop2008}{population in 2008}
  \item{pop2009}{population in 2009}
}}
\source{
\url{https://www.census.gov/popest/data/metro/totals/2009/tables/CBSA-EST2009-01.csv}
}
\description{
This data set contains population data for muSAs from 2000 to 2009 in
panel data format
}
