% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauchen.R
\name{tauchen}
\alias{tauchen}
\title{Tauchen Discretization of an AR(1)}
\usage{
tauchen(myGrid, mu, sigma)
}
\arguments{
\item{myGrid}{A vector with the equidistant support of the stochastic process}

\item{mu}{The mean of the underlying normal distribution}

\item{sigma}{The standard deviation of the underlying normal distribution}
}
\value{
a square transition probability matrix with dimension length(myGrid)
times length(myGrid).
}
\description{
Tauchen computes the transition probability matrix for a discretized
AR(1) with normally distributed innovations
}
\examples{
tauchen( c(1,2,3), 0, 100)
}
