% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{valuefunction}
\alias{valuefunction}
\title{Compute equilibrium value functions}
\usage{
valuefunction(Settings, Param)
}
\arguments{
\item{Settings}{is an R list of settings}

\item{Param}{is an R list of of parameters}
}
\value{
List with value functions, entry probabilities and
certain survival probabilities.
}
\description{
This function computes the equilibrium value functions
and the probabilities of entry and certain survival. Value
functions are computed via backward recursion. For this
function to work, the following \code{Settings} must have been
set:
\itemize{
 \item{\code{nCheck} The maximum number of firms with non-zero flow profits in some state of the world.}
 \item{\code{logGrid} The log of the grid}
 \item{\code{tolInner} The convergence tolerance for the inner loop}
 \item{\code{maxIterInner} The maximum number of iterations for the inner loop}
}
The following elements in \code{Param} must have been set
\itemize{
 \item{\code{k} A vector of length \code{nCheck} that determines the flow surplus}
 \item{\code{phi} A vector of length \code{nCheck} that governs the entry costs}
 \item{\code{omega} A scalar that governs the variance of the cost shocks}
 \item{\code{rho} The discount factor}
 \item{\code{Pi} A transition probability matrix for the demand grid, which is computed via
 \code{tauchen}}
}
}
